# Conflicts:
#   client/app/page.tsx

+++ server/src/main/java/com/b110/jjeonchongmu/domain/schedule/service/ScheduleService.java
@@ -1,5 +1,7 @@
 package com.b110.jjeonchongmu.domain.schedule.service;
 
+import com.b110.jjeonchongmu.domain.account.entity.ScheduleAccount;
+import com.b110.jjeonchongmu.domain.gathering.repo.GatheringMemberRepo;
 import com.b110.jjeonchongmu.domain.gathering.repo.GatheringRepo;
 import com.b110.jjeonchongmu.domain.schedule.dto.*;
 import com.b110.jjeonchongmu.domain.schedule.entity.Schedule;
@@ -13,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.lang.reflect.Array;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -24,18 +27,23 @@ import com.b110.jjeonchongmu.domain.gathering.entity.Gathering;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
+//@Transactional
 public class ScheduleService {
 
     private final ScheduleRepo scheduleRepo;
     private final GatheringRepo gatheringRepo;
     private final UserRepo userRepo;
     private final ScheduleMemberRepo scheduleMemberRepo;
-
+    private final GatheringMemberRepo gatheringMemberRepo;
 
 
     // 모임 일정목록 조회
     public List<ScheduleDTO> getScheduleList(Long userId, Long gatheringId) {
+        boolean isMember = gatheringMemberRepo.existsByUserIdAndGatheringId(userId, gatheringId);
+        if (!isMember) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User does not have access to this gathering");
+        }
+
         return scheduleRepo.findByGatheringGatheringId(gatheringId)
                 .stream()
                 .map(ScheduleDTO::from)
@@ -44,27 +52,30 @@ public class ScheduleService {
 
     // 일정 상세조회
     public ScheduleDetailDTO getScheduleDetail(Long userId, Long scheduleId) {
+        boolean isMember = scheduleMemberRepo.existsByUserIdAndGatheringId(userId, scheduleId);
+        if (!isMember) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User does not have access to this schedule");
+        }
+
         Schedule schedule = scheduleRepo.findById(scheduleId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Schedule not found"));
         return ScheduleDetailDTO.from(schedule);
     }
 
     // 일정 생성(총무만)
-    public void createSchedule(Long userId,Long gatheringId ,ScheduleCreateDTO scheduleCreateDTO) {
+    public Long createSchedule(Long userId,Long gatheringId ,ScheduleCreateDTO scheduleCreateDTO) {
+        boolean isManager = gatheringRepo.existsByUserIdAndGatheringId(userId, gatheringId);
+        if (!isManager) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User does not have access to this schedule");
+        }
+
         Gathering gathering = gatheringRepo.findById(gatheringId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Gathering not found"));
 
         User subManager = userRepo.findByUserId(userId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
 
-
-//        private ScheduleAccount scheduleAccount;
-//        일정 계좌번호 받아오는 로직 필요
-//        private List<ScheduleMember> attendees;
-//        일정 멤버들 받아오는 로직 필요
-
-//        생성과 동시에 계좌를 생성해야하고 총무, 부총무가 일정멤버로 지정되어야 함
-
+//        일정 생성후에 계좌를 생성해야하고  부총무가 일정멤버로 지정되어야 함
         Schedule schedule = Schedule.builder()
                 .gathering(gathering)
                 .subManager(subManager)
@@ -78,29 +89,45 @@ public class ScheduleService {
                 .status(0)
                 .build();
 
-        scheduleRepo.save(schedule);
+        Schedule savedSchedule = scheduleRepo.save(schedule);
 
+        // 부총무 일정멤버에 추가
+        List<ScheduleMember> attendees = new ArrayList<>();
+        ScheduleMember attendee = ScheduleMember.builder()
+                        .schedule(savedSchedule)
+                        .scheduleMember(subManager)
+                        .build();
+        attendees.add(attendee);
+
+        scheduleRepo.save(savedSchedule);
+
+        return savedSchedule.getId();
     }
-    // 일정 수정(총무만)
+    // 일정 수정(부총무만)
     public void updateSchedule(Long userId,Long scheduleId,ScheduleUpdateDTO scheduleUpdateDTO) {
+        boolean isSubManager = scheduleRepo.checkExistsByUserIdAndScheduleId(userId, scheduleId);
+        if (!isSubManager) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User does not have access to this schedule");
+        }
+
+
         Schedule schedule = scheduleRepo.findById(scheduleId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "schedule not found"));
 
         schedule.updateTitle(scheduleUpdateDTO.getScheduleTitle());
-        schedule.updateTitle(scheduleUpdateDTO.getScheduleTitle());
-        schedule.updateTitle(scheduleUpdateDTO.getScheduleTitle());
-        schedule.updateTitle(scheduleUpdateDTO.getScheduleTitle());
+        schedule.updateDetail(scheduleUpdateDTO.getScheduleDetail());
+        schedule.updatePlace(scheduleUpdateDTO.getSchedulePlace());
+        schedule.updatePerBudget(scheduleUpdateDTO.getPerBudget());
 
     }
-    // 일정 삭제(총무만)
+    // 일정 삭제(부총무만)
     public void deleteSchedule(Long userId, Long scheduleId) {
-//        Schedule schedule = scheduleRepo.findById(scheduleId)
-//                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Schedule not found"));
-//
-//        // 삭제 권한 확인 (옵션)
-//        if (!schedule.getSubManager().getUserId().equals(userId)) {
-//            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only submanager can delete schedule");
-//        }
+        scheduleRepo.findById(scheduleId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Schedule not found"));
+        boolean isSubManager = scheduleRepo.checkExistsByUserIdAndScheduleId(userId, scheduleId);
+        if (!isSubManager) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User does not have access to this schedule");
+        }
         
         scheduleRepo.deleteById(scheduleId);
     }
